refactor: create body component for quiz page
