+++ task/src/pages/Quiz/Quiz.module.scss
@@ -3,256 +3,9 @@
   min-height: 100vh;
   width: 100%;
 
-  &-img {
-    background-repeat: no-repeat;
-    background-size: contain;
-    border-radius: 10px;
-    height: 155px;
-    margin: 0;
-    width: 200px;
-  }
-
-  &__preview {
-    background-color: #303030;
-    border-radius: 0.25rem;
-    color: #fff;
-    display: flex;
-    margin-bottom: 25px;
-    padding: 1rem;
-  }
-
-  &__preview-img {
-    background-image: url('../../assets/bird.06a46938.jpg');
-    background-repeat: no-repeat;
-    background-size: contain;
-    border-radius: 10px;
-    height: 155px;
-    margin: 0;
-    width: 200px;
-  }
-
-  &__preview_question {
-    display: flex;
-    flex-direction: column;
-    margin-bottom: 0;
-    margin-left: 40px;
-    padding-left: 0;
-    width: calc(100% - 240px);
-  }
-
-  &__preview_question-header {
-    background-color: #303030;
-    border: 1px solid #555;
-    border-left-width: 0;
-    border-radius: 0;
-    border-right-width: 0;
-    border-top-width: 0;
-    font-size: 2rem;
-    padding: 0.25rem;
-
-  }
-
-  &-audio {
-    margin: 10px 40px 8px 0;
-    min-width: 100px;
-    outline: 0;
-    padding: 0.25rem;
-    width: 100%;
-  }
-
   &-container {
-    //display: flex;
     margin: 0 auto;
     max-width: 1024px;
     padding: 0 40px;
-  }
-
-  &__interactive {
-    display: flex;
-    justify-content: space-between;
-  }
-
-  &__questions {
-    display: flex;
-  }
-
-  &__questions-container {
-    color: #fff;
-    display: flex;
-    flex-direction: column;
-    width: 48%;
-  }
-
-  &__questions-question {
-    align-items: center;
-    background-color: #303030;
-    border: 0.25px solid #555;
-    border-left-width: 0.5px;
-    border-right-width: 0.5px;
-    display: flex;
-    padding: 0.85rem 1.25rem;
-    transition: 0.3s;
-  }
-
-  &__questions-question:nth-child(1) {
-    border-top-left-radius: 0.25rem;
-    border-top-right-radius: 0.25rem;
-  }
-
-  &__questions-question:last-child {
-    border-bottom-left-radius: 0.25rem;
-    border-bottom-right-radius: 0.25rem;
-  }
-
-  &__questions-question:hover {
-    background-color: rgb(72, 72, 72);
-    transition: 0.3s;
-  }
-
-  &__questions-indicator {
-    border-radius: 50%;
-    display: inline-block;
-    height: 10px;
-    margin-right: 15px;
-    transition: 0.1s;
-    width: 10px;
-  }
-
-  &__questions-indicator--inactive {
-    background-color: #666;
-  }
-
-  &__questions-indicator--false {
-    background-color: #d62c1a;
-  }
-
-  &__questions-indicator--true {
-    background-color: #00bc8c;
-  }
-
-  &__instruction {
-    align-items: center;
-    background-color: #303030;
-    border: 0.25px solid #555;
-    border-left-width: 0.5px;
-    border-radius: 0.25rem;
-    border-right-width: 0.5px;
-    color: #fff;
-    display: flex;
-    flex-direction: column;
-    transition: 0.3s;
-    width: calc(48% - 1rem);
-  }
-
-  &__instruction-container {
-    padding: 1rem;
-  }
-
-  &__instruction_body {
-    display: flex;
-    flex-wrap: wrap;
-    min-height: 173px;
-    width: 100%;
-  }
-
-  &__instruction_body-img {
-    border-radius: 10px;
-    height: 155px;
-    margin-right: 2rem;
-    object-fit: cover;
-    width: 200px;
-  }
-
-  &__instruction_body-data {
-    width: 100%;
-  }
-
-  &__instruction_body-name {
-    border-bottom: 1px solid #555;
-    font-size: 1.40625rem;
-    font-weight: 500;
-    line-height: 1.2;
-    margin-bottom: 0.5rem;
-    padding: 0.25rem;
-  }
-
-  &__instruction_body-species {
-    border-bottom: 1px solid #555;
-    margin-bottom: 0.5rem;
-    padding: 0.25rem;
-  }
-
-  &__endgame {
-    background-color: #303030;
-    border-radius: 0.3rem;
-    color: #fff;
-    display: flex;
-    flex-direction: column;
-    margin-bottom: 2rem;
-    padding: 4rem 2rem;
-    text-align: center;
-  }
-
-  &__endgame-Header {
-    font-size: 4.5rem;
-    font-weight: 300;
-    line-height: 1.2;
-  }
-
-  &__next {
-    background-color: #00bc8c;
-    border-radius: 8px;
-    color: #fff;
-    margin-bottom: 30px;
-    margin-top: 30px;
-    padding: 10px;
-    transition: 0.3s;
-    width: 100%;
-  }
-
-  &__next:hover {
-    background-color: #00e4b0;
-    transition: 0.3s;
-  }
-}
-
-@media (max-width: 700px) {
-  .Quiz {
-    &__interactive {
-      align-items: center;
-      flex-direction: column;
-    }
-
-    &__questions {
-      display: flex;
-    }
-
-    &__questions-container {
-      margin-bottom: 2rem;
-      width: 100%;
-    }
-
-    &__instruction {
-      width: 100%;
-    }
-  }
-}
-
-@media (max-width: 550px) {
-  .Quiz {
-    &__preview {
-      align-items: center;
-      flex-direction: column;
-    }
-
-    &__preview_question {
-      margin: 20px 0 0;
-      text-align: center;
-      width: 100%;
-    }
-
-    &__preview &-audio {
-      width: 100%;
-    }
   }
 }
